+++ www/zmq/publisher_block_info.go
@@ -26,7 +26,7 @@ func (b *blockInfoPub) onNewBlock(blk *block.Block) {
 		blk.Header().ProposerAddress(),
 		blk.Header().UnixTime(),
 		uint16(len(blk.Transactions())),
-		blk.Height(),
+		uint32(blk.Height()),
 	)
 
 	message := zmq4.NewMsg(rawMsg)
